+++ files/default/tests/minitest/default_test.rb
@@ -1,7 +1,7 @@
 require File.expand_path('../support/helpers', __FILE__)
 
-describe_recipe "template-cookbook::default" do
-  include Helpers::TemplateCookbook
+describe_recipe "mysql_connector::default" do
+  include Helpers::MysqlConnector
 
   # do work, son.
 end