+++ files/default/tests/minitest/support/helpers.rb
@@ -1,13 +1,7 @@
 module Helpers
-  module TemplateCookbook
-    require 'chef/mixin/shell_out'
-    include Chef::Mixin::ShellOut
+  module MysqlConnector
     include MiniTest::Chef::Assertions
     include MiniTest::Chef::Context
     include MiniTest::Chef::Resources
-
-    def ran_recipe?(recipe)
-      node.run_state[:seen_recipes].keys.include?(recipe)
-    end
   end
 end