+++ metadata.rb
@@ -1,16 +1,16 @@
-name              "template-cookbook"
-maintainer        "The Wharton School - The University of Pennsylvania"
-maintainer_email  "chef-admins@wharton.upenn.edu"
+name              "mysql_connector"
+maintainer        "Brian Flad"
+maintainer_email  "bflad@wharton.upenn.edu"
 license           "Apache 2.0"
-description       "A cool description."
+description       "Recipes/LWRPs for installing MySQL Connectors."
 version           "0.1.0"
-recipe            "template-cookbook", "Short description of template-cookbook default recipe."
-recipe            "template-cookbook::alternate", "Short description of template-cookbook::alternate recipe."
+recipe            "mysql_connector", "Empty recipe for including LWRPs."
+recipe            "mysql_connector::odbc_package", "Installs MySQL Connector/ODBC via package."
 
-%w{ apache2 logrotate }.each do |d|
+%w{ ark }.each do |d|
   depends d
 end
 
-%w{ redhat ubuntu }.each do |os|
+%w{ redhat }.each do |os|
   supports os
 end