+++ providers/j.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: mysql_connector
-# Recipe:: default
+# Provider:: j
 #
 # Copyright 2012
 #
@@ -17,4 +17,12 @@
 # limitations under the License.
 #
 
-# This recipe intentionally blank for including LWRPs in your cookbook.
+action :create do
+  ark "mysql-connector-java" do
+    url node['mysql_connector']['j']['url']
+    creates "mysql-connector-java-#{node['mysql_connector']['j']['version']}-bin.jar"
+    path new_resource.path
+    action :cherry_pick
+  end
+  new_resource.updated_by_last_action(true)
+end