+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## v0.1.0
 
-* Initial release.
+* Initial release with MySQL Connector/J LWRP.